+++ core/src/plugins/kubernetes/kubernetes-type/common.ts
@@ -9,13 +9,13 @@
 import { resolve } from "path"
 import { readFile } from "fs-extra"
 import Bluebird from "bluebird"
-import { flatten, keyBy, set } from "lodash"
+import { flatten, set } from "lodash"
 import { loadAll } from "js-yaml"
 
 import { KubernetesModule } from "./module-config"
 import { KubernetesResource } from "../types"
 import { KubeApi } from "../api"
-import { gardenAnnotationKey, stableStringify } from "../../../util/string"
+import { gardenAnnotationKey } from "../../../util/string"
 import { Log } from "../../../logger/log-entry"
 import { PluginContext } from "../../../plugin-context"
 import { ConfigurationError, PluginError } from "../../../exceptions"
@@ -24,10 +24,9 @@ import { HelmModule } from "../helm/module-config"
 import { KubernetesDeployAction } from "./config"
 import { CommonRunParams } from "../../../plugin/handlers/Run/run"
 import { runAndCopy } from "../run"
-import { getTargetResource, getResourcePodSpec, getResourceContainer, makePodName, getResourceKey } from "../util"
-import { ActionMode, Resolved } from "../../../actions/types"
+import { getTargetResource, getResourcePodSpec, getResourceContainer, makePodName } from "../util"
+import { Resolved } from "../../../actions/types"
 import { KubernetesPodRunAction, KubernetesPodTestAction } from "./kubernetes-pod"
-import { V1ConfigMap } from "@kubernetes/client-node"
 import { glob } from "glob"
 
 /**
@@ -40,14 +39,16 @@ export async function getManifests({
   log,
   action,
   defaultNamespace,
+  readFromSrcDir = false,
 }: {
   ctx: PluginContext
   api: KubeApi
   log: Log
   action: Resolved<KubernetesDeployAction | KubernetesPodRunAction | KubernetesPodTestAction>
   defaultNamespace: string
+  readFromSrcDir?: boolean
 }): Promise<KubernetesResource[]> {
-  const rawManifests = (await readManifests(ctx, action, log)) as KubernetesResource[]
+  const rawManifests = (await readManifests(ctx, action, log, readFromSrcDir)) as KubernetesResource[]
 
   // remove *List objects
   const manifests = rawManifests.flatMap((manifest) => {
@@ -69,11 +70,6 @@ export async function getManifests({
     }
     return manifest
   })
-
-  if (action.kind === "Deploy") {
-    // Add metadata ConfigMap to aid quick status check
-    manifests.push(getMetadataManifest(action, defaultNamespace, manifests))
-  }
 
   return Bluebird.map(manifests, async (manifest) => {
     // Ensure a namespace is set, if not already set, and if required by the resource type
@@ -110,80 +106,26 @@ export async function getManifests({
   })
 }
 
-export interface ManifestMetadata {
-  key: string
-  apiVersion: string
-  kind: string
-  name: string
-  namespace: string
-}
-
-export interface ParsedMetadataManifestData {
-  resolvedVersion: string
-  mode: ActionMode
-  manifestMetadata: { [key: string]: ManifestMetadata }
-}
-
-export function getMetadataManifest(
-  action: Resolved<KubernetesDeployAction>,
-  defaultNamespace: string,
-  manifests: KubernetesResource[]
-): KubernetesResource<V1ConfigMap> {
-  const manifestMetadata: ManifestMetadata[] = manifests.map((m) => ({
-    key: getResourceKey(m),
-    apiVersion: m.apiVersion,
-    kind: m.kind,
-    name: m.metadata.name,
-    namespace: m.metadata.namespace || defaultNamespace,
-  }))
-
-  return {
-    apiVersion: "v1",
-    kind: "ConfigMap",
-    metadata: {
-      name: `garden-meta-${action.kind.toLowerCase()}-${action.name}`,
-    },
-    data: {
-      resolvedVersion: action.versionString(),
-      mode: action.mode(),
-      manifestMetadata: stableStringify(keyBy(manifestMetadata, "key")),
-    },
-  }
-}
-
-export function parseMetadataResource(log: Log, resource: KubernetesResource<V1ConfigMap>): ParsedMetadataManifestData {
-  // TODO: validate schema here
-  const output: ParsedMetadataManifestData = {
-    resolvedVersion: resource.data?.resolvedVersion || "",
-    mode: (resource.data?.mode || "default") as ActionMode,
-    manifestMetadata: {},
-  }
-
-  const manifestMetadata = resource.data?.manifestMetadata
-
-  if (manifestMetadata) {
-    try {
-      // TODO: validate by schema
-      output.manifestMetadata = JSON.parse(manifestMetadata)
-    } catch (error) {
-      log.debug({ msg: `Failed querying for remote resources: ${error.message}`, error })
-    }
-  }
-
-  return output
-}
-
 const disallowedKustomizeArgs = ["-o", "--output", "-h", "--help"]
 
 /**
  * Read the manifests from the module config, as well as any referenced files in the config.
+ *
+ * @param module The kubernetes module to read manifests for.
+ * @param readFromSrcDir Whether or not to read the manifests from the module build dir or from the module source dir.
+ * In general we want to read from the build dir to ensure that manifests added via the `build.dependencies[].copy`
+ * field will be included. However, in some cases, e.g. when getting the service status, we can't be certain that
+ * the build has been staged and we therefore read the manifests from the source.
+ *
+ * TODO: Remove this once we're checking for kubernetes module service statuses with version hashes.
  */
 export async function readManifests(
   ctx: PluginContext,
   action: Resolved<KubernetesDeployAction | KubernetesPodRunAction | KubernetesPodTestAction>,
-  log: Log
+  log: Log,
+  readFromSrcDir = false
 ) {
-  const manifestPath = action.getBuildPath()
+  const manifestPath = readFromSrcDir ? action.basePath() : action.getBuildPath()
 
   const spec = action.getSpec()